requirements.txt
# frontend
dash==2.18.2
dash-bootstrap-components==1.7.1
plotly==6.0.0

# llms
google-genai==1.3.0 # for gemini

# environment
python-dotenv==1.0.1

# google connections
google-api-python-client==2.162.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
google-cloud-storage==2.19.0
google-cloud-aiplatform==1.85.0
google-cloud-storage-control==1.3.1

weaviate-client==4.11.3
unstructured[all-docs]==0.17.2
extract-msg==0.53.1
python-docx==1.1.2
pandas==2.2.3
pillow==11.1.0
tiktoken==0.9.0
yfinance==0.2.55
markdown-pdf==1.6